cleaning up mite (#240)

+++ mite/recorder.py
@@ -1,3 +1,4 @@
+import contextlib
 import logging
 import os
 
@@ -12,21 +13,16 @@ class Recorder:
         if target_dir is None:
             target_dir = "recorded_data"
         self._target_dir = os.path.abspath(target_dir)
-        try:
+        with contextlib.suppress(FileExistsError):
             os.makedirs(self._target_dir)
-        except FileExistsError:
-            pass
 
     def process_message(self, msg):
-        if "type" in msg and "name" in msg:
-            if msg["name"] is not None:
-                msg_file = os.path.join(self._target_dir, msg["name"] + ".msgpack")
-                if msg["type"] == "data_created":
-                    open(msg_file, "ab").write(pack_msg(msg["data"]))
-                elif msg["type"] == "purge_data":
-                    try:
-                        os.remove(msg_file)
-                    except FileNotFoundError:
-                        logger.info(
-                            "Tried to delete a file that didn't exist, continuing"
-                        )
+        if "type" in msg and "name" in msg and msg["name"] is not None:
+            msg_file = os.path.join(self._target_dir, msg["name"] + ".msgpack")
+            if msg["type"] == "data_created":
+                open(msg_file, "ab").write(pack_msg(msg["data"]))
+            elif msg["type"] == "purge_data":
+                try:
+                    os.remove(msg_file)
+                except FileNotFoundError:
+                    logger.info("Tried to delete a file that didn't exist, continuing")
